// src/api/model/apiBeycoCoffeeQuality.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * Abelium INATrace Services API swagger documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */





export interface ApiBeycoCoffeeQuality { 
    /**
     * Quality of coffee beans
     */
    type?: ApiBeycoCoffeeQuality.TypeEnum;
}

/**
 * Namespace for property- and property-value-enumerations of ApiBeycoCoffeeQuality.
 */
export namespace ApiBeycoCoffeeQuality {
    /**
     * All properties of ApiBeycoCoffeeQuality.
     */
    export enum Properties {
        /**
         * Quality of coffee beans
         */
        type = 'type'
    }

    /**
     * All possible values of type.
     */
    export enum TypeEnum {
        Specialty = 'Specialty',
        CommodityConventional = 'CommodityConventional',
        LowGrade = 'LowGrade'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiBeycoCoffeeQuality',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: true,
                    datatypeWithEnum: 'ApiBeycoCoffeeQuality.TypeEnum',
                    required: false,
                    name: 'type',
                    classname: 'ApiBeycoCoffeeQuality',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                type: [
                ],
            }
        }
    }

  // export const ApiBeycoCoffeeQualityValidationScheme = {
  //     validators: [],
  //     fields: {
  //               type: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiBeycoCoffeeQuality>;


}