// src/api/model/apiBeycoCoffeeCertificate.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * Abelium INATrace Services API swagger documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */





export interface ApiBeycoCoffeeCertificate { 
    /**
     * Certificate of coffee beans
     */
    type?: ApiBeycoCoffeeCertificate.TypeEnum;
}

/**
 * Namespace for property- and property-value-enumerations of ApiBeycoCoffeeCertificate.
 */
export namespace ApiBeycoCoffeeCertificate {
    /**
     * All properties of ApiBeycoCoffeeCertificate.
     */
    export enum Properties {
        /**
         * Certificate of coffee beans
         */
        type = 'type'
    }

    /**
     * All possible values of type.
     */
    export enum TypeEnum {
        Ra = 'Ra',
        Fairtrade = 'Fairtrade',
        Organic = 'Organic',
        Practices = 'Practices',
        FourC = 'FourC'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiBeycoCoffeeCertificate',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: true,
                    datatypeWithEnum: 'ApiBeycoCoffeeCertificate.TypeEnum',
                    required: false,
                    name: 'type',
                    classname: 'ApiBeycoCoffeeCertificate',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                type: [
                ],
            }
        }
    }

  // export const ApiBeycoCoffeeCertificateValidationScheme = {
  //     validators: [],
  //     fields: {
  //               type: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiBeycoCoffeeCertificate>;


}